+++ app/src/main/java/com/inverse/unofficial/proxertv/model/ServerConfig.kt
@@ -1,8 +1,9 @@
 package com.inverse.unofficial.proxertv.model
 
 class ServerConfig(
-        schema: String = "https",
-        val host: String = "proxer.me") {
+    schema: String = "https",
+    val host: String = "proxer.me"
+) {
     val baseUrl = "$schema://$host/"
     val apiBaseUrl = baseUrl + "api/v1/"
     val updatesListUrl = baseUrl + "anime/updates#top"
@@ -13,10 +14,11 @@ class ServerConfig(
     fun topRatingMovieListUrl(page: Int = 1) = baseUrl + "anime/movie/rating/all/$page#top"
 
     fun searchUrl(query: String) = baseUrl + "search?s=search&name=$query&typ=all-anime"
-    fun episodeStreamsUrl(seriesId: Int, episodeNum: Int, subType: String) = baseUrl + "watch/$seriesId/$episodeNum/$subType"
+    fun episodeStreamsUrl(seriesId: Long, episodeNum: Int, subType: String) =
+        baseUrl + "watch/$seriesId/$episodeNum/$subType"
 
     companion object {
         // for ease of use
-        fun coverUrl(seriesId: Int) = "https://cdn.proxer.me/cover/$seriesId.jpg"
+        fun coverUrl(seriesId: Long) = "https://cdn.proxer.me/cover/$seriesId.jpg"
     }
 }
